process_data: Now checking if MTS files exist before processing

+++ processing/process_data.sh
@@ -170,43 +170,48 @@ sct_process_segmentation -i ${file_t2_seg}.nii.gz -vert 2:3 -vertfile PAM50_leve
 file_t1w="${SUBJECT}_acq-T1w_MTS"
 file_mton="${SUBJECT}_acq-MTon_MTS"
 file_mtoff="${SUBJECT}_acq-MToff_MTS"
-# Fetch TR and FA from the json files
-FA_t1w=$(get_field_from_json ${file_t1w}.json FlipAngle)
-FA_mton=$(get_field_from_json ${file_mton}.json FlipAngle)
-FA_mtoff=$(get_field_from_json ${file_mtoff}.json FlipAngle)
-TR_t1w=$(get_field_from_json ${file_t1w}.json RepetitionTime)
-TR_mton=$(get_field_from_json ${file_mton}.json RepetitionTime)
-TR_mtoff=$(get_field_from_json ${file_mtoff}.json RepetitionTime)
-# Segment spinal cord (only if it does not exist)
-segment_if_does_not_exist $file_t1w "t1"
-file_t1w_seg=$FILESEG
-# Create mask
-sct_create_mask -i ${file_t1w}.nii.gz -p centerline,${file_t1w_seg}.nii.gz -size 35mm -o ${file_t1w}_mask.nii.gz
-# Crop data for faster processing
-sct_crop_image -i ${file_t1w}.nii.gz -m ${file_t1w}_mask.nii.gz -o ${file_t1w}_crop.nii.gz
-file_t1w="${file_t1w}_crop"
-# Register PD->T1w
-# Tips: here we only use rigid transformation because both images have very similar sequence parameters. We don't want to use SyN/BSplineSyN to avoid introducing spurious deformations.
-sct_register_multimodal -i ${file_mtoff}.nii.gz -d ${file_t1w}.nii.gz -param step=1,type=im,algo=rigid,slicewise=1,metric=CC -x spline
-file_mtoff="${file_mtoff}_reg"
-# Register MT->T1w
-sct_register_multimodal -i ${file_mton}.nii.gz -d ${file_t1w}.nii.gz -param step=1,type=im,algo=rigid,slicewise=1,metric=CC -x spline
-file_mton="${file_mton}_reg"
-# Register template->T1w_ax (using template-T1w as initial transformation)
-sct_register_multimodal -i $SCT_DIR/data/PAM50/template/PAM50_t1.nii.gz -iseg $SCT_DIR/data/PAM50/template/PAM50_cord.nii.gz -d ${file_t1w}.nii.gz -dseg ${file_t1w_seg}.nii.gz -param step=1,type=seg,algo=slicereg,metric=MeanSquares,smooth=2:step=2,type=im,algo=syn,metric=CC,iter=5,gradStep=0.5 -initwarp warp_template2T1w.nii.gz -initwarpinv warp_T1w2template.nii.gz
-# Rename warping field for clarity
-mv warp_PAM50_t12${file_t1w}.nii.gz warp_template2axT1w.nii.gz
-mv warp_${file_t1w}2PAM50_t1.nii.gz warp_axT1w2template.nii.gz
-# Warp template
-sct_warp_template -d ${file_t1w}.nii.gz -w warp_template2axT1w.nii.gz -ofolder label_axT1w -qc ${PATH_QC} -qc-subject ${SUBJECT}
-# Compute MTR
-sct_compute_mtr -mt0 ${file_mtoff}.nii.gz -mt1 ${file_mton}.nii.gz
-# Compute MTsat
-sct_compute_mtsat -mt ${file_mton}.nii.gz -pd ${file_mtoff}.nii.gz -t1 ${file_t1w}.nii.gz -trmt $TR_mton -trpd $TR_mtoff -trt1 $TR_t1w -famt $FA_mton -fapd $FA_mtoff -fat1 $FA_t1w
-# Extract MTR, MTsat and T1 in WM between C2 and C5 vertebral levels
-sct_extract_metric -i mtr.nii.gz -f label_axT1w/atlas -l 51 -vert 2:5 -vertfile label_axT1w/template/PAM50_levels.nii.gz -o ${PATH_RESULTS}/MTR.csv -append 1
-sct_extract_metric -i mtsat.nii.gz -f label_axT1w/atlas -l 51 -vert 2:5 -vertfile label_axT1w/template/PAM50_levels.nii.gz -o ${PATH_RESULTS}/MTsat.csv -append 1
-sct_extract_metric -i t1map.nii.gz -f label_axT1w/atlas -l 51 -vert 2:5 -vertfile label_axT1w/template/PAM50_levels.nii.gz -o ${PATH_RESULTS}/T1.csv -append 1
+
+if [[ -e "${file_t1w}.nii.gz" && -e "${file_mton}.nii.gz" && -e "${file_mtoff}.nii.gz" ]]; then
+  # Fetch TR and FA from the json files
+  FA_t1w=$(get_field_from_json ${file_t1w}.json FlipAngle)
+  FA_mton=$(get_field_from_json ${file_mton}.json FlipAngle)
+  FA_mtoff=$(get_field_from_json ${file_mtoff}.json FlipAngle)
+  TR_t1w=$(get_field_from_json ${file_t1w}.json RepetitionTime)
+  TR_mton=$(get_field_from_json ${file_mton}.json RepetitionTime)
+  TR_mtoff=$(get_field_from_json ${file_mtoff}.json RepetitionTime)
+  # Segment spinal cord (only if it does not exist)
+  segment_if_does_not_exist $file_t1w "t1"
+  file_t1w_seg=$FILESEG
+  # Create mask
+  sct_create_mask -i ${file_t1w}.nii.gz -p centerline,${file_t1w_seg}.nii.gz -size 35mm -o ${file_t1w}_mask.nii.gz
+  # Crop data for faster processing
+  sct_crop_image -i ${file_t1w}.nii.gz -m ${file_t1w}_mask.nii.gz -o ${file_t1w}_crop.nii.gz
+  file_t1w="${file_t1w}_crop"
+  # Register PD->T1w
+  # Tips: here we only use rigid transformation because both images have very similar sequence parameters. We don't want to use SyN/BSplineSyN to avoid introducing spurious deformations.
+  sct_register_multimodal -i ${file_mtoff}.nii.gz -d ${file_t1w}.nii.gz -param step=1,type=im,algo=rigid,slicewise=1,metric=CC -x spline
+  file_mtoff="${file_mtoff}_reg"
+  # Register MT->T1w
+  sct_register_multimodal -i ${file_mton}.nii.gz -d ${file_t1w}.nii.gz -param step=1,type=im,algo=rigid,slicewise=1,metric=CC -x spline
+  file_mton="${file_mton}_reg"
+  # Register template->T1w_ax (using template-T1w as initial transformation)
+  sct_register_multimodal -i $SCT_DIR/data/PAM50/template/PAM50_t1.nii.gz -iseg $SCT_DIR/data/PAM50/template/PAM50_cord.nii.gz -d ${file_t1w}.nii.gz -dseg ${file_t1w_seg}.nii.gz -param step=1,type=seg,algo=slicereg,metric=MeanSquares,smooth=2:step=2,type=im,algo=syn,metric=CC,iter=5,gradStep=0.5 -initwarp warp_template2T1w.nii.gz -initwarpinv warp_T1w2template.nii.gz
+  # Rename warping field for clarity
+  mv warp_PAM50_t12${file_t1w}.nii.gz warp_template2axT1w.nii.gz
+  mv warp_${file_t1w}2PAM50_t1.nii.gz warp_axT1w2template.nii.gz
+  # Warp template
+  sct_warp_template -d ${file_t1w}.nii.gz -w warp_template2axT1w.nii.gz -ofolder label_axT1w -qc ${PATH_QC} -qc-subject ${SUBJECT}
+  # Compute MTR
+  sct_compute_mtr -mt0 ${file_mtoff}.nii.gz -mt1 ${file_mton}.nii.gz
+  # Compute MTsat
+  sct_compute_mtsat -mt ${file_mton}.nii.gz -pd ${file_mtoff}.nii.gz -t1 ${file_t1w}.nii.gz -trmt $TR_mton -trpd $TR_mtoff -trt1 $TR_t1w -famt $FA_mton -fapd $FA_mtoff -fat1 $FA_t1w
+  # Extract MTR, MTsat and T1 in WM between C2 and C5 vertebral levels
+  sct_extract_metric -i mtr.nii.gz -f label_axT1w/atlas -l 51 -vert 2:5 -vertfile label_axT1w/template/PAM50_levels.nii.gz -o ${PATH_RESULTS}/MTR.csv -append 1
+  sct_extract_metric -i mtsat.nii.gz -f label_axT1w/atlas -l 51 -vert 2:5 -vertfile label_axT1w/template/PAM50_levels.nii.gz -o ${PATH_RESULTS}/MTsat.csv -append 1
+  sct_extract_metric -i t1map.nii.gz -f label_axT1w/atlas -l 51 -vert 2:5 -vertfile label_axT1w/template/PAM50_levels.nii.gz -o ${PATH_RESULTS}/T1.csv -append 1
+else
+  echo "WARNING: MTS dataset is incomplete."
+fi
 
 # t2s
 # ------------------------------------------------------------------------------
